HSEARCH-3729 Split MassIndexingFailureIT
... so that we can test a custom MassIndexingFailureHandler more easily.
